requirements.txt
nltk==3.8.1
sklearn-pandas==2.2.0
scikit-learn==1.2.2
numpy==1.23.5
scikit-learn==1.2.2